test: split remote process fixtures
