+++ Packages/macOS/CmuxRemoteSession/Tests/CmuxRemoteSessionTests/ProcessRunEvent.swift
@@ -1,0 +1,7 @@
+import Foundation
+@testable import CmuxRemoteSession
+
+enum ProcessRunEvent: Sendable {
+    case completed(Result<RemoteCommandResult, any Error>)
+    case deadline
+}
